:pencil2: Rename TypeScript Files
