+++ typescript/src/IExtEuclid.ts
